Feat/edit rate and review for tour api (#61)

+++ src/main/java/com/example/localguidebe/repository/ReviewRepository.java
@@ -1,12 +1,11 @@
 package com.example.localguidebe.repository;
 
 import com.example.localguidebe.entity.Review;
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
-public interface ReviewRepository extends JpaRepository<Review,Integer>{
-    List<Review> getReviewsByGuideId(Long guideId);
+public interface ReviewRepository extends JpaRepository<Review, Long> {
+  List<Review> getReviewsByGuideId(Long guideId);
 }
